+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/properties/CardinalitiesProperty.java
@@ -27,6 +27,7 @@ import com.apple.foundationdb.record.query.plan.cascades.ExpressionRef;
 import com.apple.foundationdb.record.query.plan.cascades.Quantifier;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.ExplodeExpression;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.FullUnorderedScanExpression;
+import com.apple.foundationdb.record.query.plan.cascades.expressions.GroupByExpression;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.LogicalDistinctExpression;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.LogicalFilterExpression;
 import com.apple.foundationdb.record.query.plan.cascades.expressions.LogicalIntersectionExpression;
@@ -271,6 +272,15 @@ public class CardinalitiesProperty implements ExpressionProperty<CardinalitiesPr
     @Nonnull
     @Override
     public Cardinalities visitRecordQueryStreamingAggregationPlan(@Nonnull final RecordQueryStreamingAggregationPlan element) {
+        // if we do not have any grouping value, we will apply the aggregation(s) over the entire child result set
+        // and return a single row comprising the aggregation(s) result
+        if (element.getGroupingValue() == null) {
+            return new Cardinalities(Cardinality.ofCardinality(1L), Cardinality.ofCardinality(1L));
+        }
+        // if the grouping value is constant, the cardinality ranges between 0 and 1.
+        if (element.getGroupingValue().isConstant()) {
+            return new Cardinalities(Cardinality.ofCardinality(0L), Cardinality.ofCardinality(1L));
+        }
         return Cardinalities.unknownCardinalities();
     }
 
@@ -365,6 +375,21 @@ public class CardinalitiesProperty implements ExpressionProperty<CardinalitiesPr
     @Nonnull
     @Override
     public Cardinalities visitFullUnorderedScanExpression(@Nonnull final FullUnorderedScanExpression element) {
+        return Cardinalities.unknownCardinalities();
+    }
+
+    @Nonnull
+    @Override
+    public Cardinalities visitGroupByExpression(@Nonnull final GroupByExpression element) {
+        // if we do not have any grouping value, we will apply the aggregation(s) over the entire child result set
+        // and return a single row comprising the aggregation(s) result
+        if (element.getGroupingValue() == null) {
+            return new Cardinalities(Cardinality.ofCardinality(1L), Cardinality.ofCardinality(1L));
+        }
+        // if the grouping value is constant, the cardinality ranges between 0 and 1.
+        if (element.getGroupingValue().isConstant()) {
+            return new Cardinalities(Cardinality.ofCardinality(0L), Cardinality.ofCardinality(1L));
+        }
         return Cardinalities.unknownCardinalities();
     }