+++ fdb-record-layer-core/src/main/java/com/apple/foundationdb/record/query/plan/cascades/properties/OrderingProperty.java
@@ -21,7 +21,6 @@
 package com.apple.foundationdb.record.query.plan.cascades.properties;
 
 import com.apple.foundationdb.annotation.API;
-import com.apple.foundationdb.record.metadata.Key;
 import com.apple.foundationdb.record.metadata.expressions.KeyExpression;
 import com.apple.foundationdb.record.query.expressions.Comparisons;
 import com.apple.foundationdb.record.query.plan.bitmap.ComposedBitmapIndexQueryPlan;
@@ -131,15 +130,8 @@ public class OrderingProperty implements PlanProperty<Ordering> {
                                 return Stream.of(Pair.of((FieldValue)valuePredicate.getValue(), valuePredicate.getComparison()));
                             })
                             .map(valueComparisonPair -> {
-
                                 final var fieldValue = valueComparisonPair.getLeft();
-                                final String fieldName = fieldValue.getFieldName();
-                                KeyExpression keyExpression =
-                                        Key.Expressions.field(fieldName);
-                                final List<String> fieldPrefix = fieldValue.getFieldPrefix();
-                                for (int i = fieldPrefix.size() - 1; i >= 0; i --) {
-                                    keyExpression = Key.Expressions.field(fieldPrefix.get(i)).nest(keyExpression);
-                                }
+                                KeyExpression keyExpression = KeyExpression.fromPath(fieldValue.getFieldPath());
                                 return Pair.of(keyExpression, valueComparisonPair.getRight());
                             })
                             .collect(ImmutableSetMultimap.toImmutableSetMultimap(Pair::getLeft, Pair::getRight));